implement authorization using Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -67,7 +67,12 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "access denied");
+        }
         return view("questions.edit", compact('question'));
+
+
     }
 
     /**
@@ -79,6 +84,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "access denied");
+        }
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', "Question updated");
     }
@@ -91,6 +99,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "access denied");
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', "Your question has been deleted");
